next_rec_two_tower/requirements-tt.txt
# Core
numpy>=1.26
pandas>=2.1
scikit-learn>=1.3

# DL & training
torch==2.2.2+cpu; platform_system != 'Linux' or platform_machine == 'x86_64'
pytorch-lightning>=2.2

# ANN
faiss-cpu>=1.7.4

# API
fastapi>=0.110
uvicorn[standard]>=0.29

# MLOps
mlflow>=2.9
optuna>=3.5
prefect>=2.14

# Data quality
great-expectations>=0.18

# Optional feature store (S3/Redis backends)
feast>=0.42
redis>=5

# Dev
pydantic>=2.6
typing-extensions>=4.9
